% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windkanal-package.R
\docType{package}
\name{windkanal-package}
\alias{windkanal}
\alias{windkanal-package}
\title{windkanal: Stochastic Simulation Testbed for Living Clinical Data Systems}
\description{
Ein Windkanal fuer lebende Versorgungs-Datensysteme: stochastische
(Monte-Carlo-)Simulation wachsender Routine-Datenstroeme
(sitzungsweise Outcomes, Patient:innen genestet in Therapeut:innen,
gestaffelter Fall-Eingang), damit Schaetzverfahren und
Analyse-Governance-Protokolle an bekannter Wahrheit geprueft werden
koennen, bevor echte Klinikdaten fliessen.
}
\details{
Die Schichten des Pakets (alle ausgeliefert; Details im README):
\enumerate{
\item Stream-Generator (\code{sim_stream()})
\item Snapshot-/Gate-Mechanik (Peeking vs. Release-Gates)
\item Estimator-Wrapper (lme4, bcf, grf) + Diagnosands
\item Kalibrierte Presets ("ambulanz_de") + Plasmode-Modus
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chrisobeser/windkanal}
  \item Report bugs at \url{https://github.com/chrisobeser/windkanal/issues}
}

}
\author{
\strong{Maintainer}: Christian Obeser \email{gersari99@gmail.com}

Authors:
\itemize{
  \item Christian Obeser \email{gersari99@gmail.com}
}

}
\keyword{internal}
